moved code to core
