+++ examples/aibs_smartspim_instrument.py
@@ -3,7 +3,7 @@
 import datetime
 
 from aind_data_schema.core.instrument import Com, Instrument
-from aind_data_schema.models.devices import (
+from aind_data_schema.core.devices import (
     AdditionalImagingDevice,
     Detector,
     Filter,
