+++ examples/exaspim_acquisition.py
@@ -4,9 +4,9 @@ from datetime import datetime, timezone
 
 from aind_data_schema.core import acquisition
 from aind_data_schema.core.procedures import Reagent
-from aind_data_schema.imaging import tile
-from aind_data_schema.models.coordinates import ImageAxis, Scale3dTransform, Translation3dTransform
-from aind_data_schema.models.devices import Calibration, Maintenance
+from aind_data_schema.core import tile
+from aind_data_schema.core.coordinates import ImageAxis, Scale3dTransform, Translation3dTransform
+from aind_data_schema.core.devices import Calibration, Maintenance
 from aind_data_schema.models.organizations import Organization
 from aind_data_schema.models.pid_names import PIDName
 from aind_data_schema.models.registry import Registry
