+++ src/aind_data_schema/core/coordinates.py
