+++ src/aind_data_schema/core/devices.py
@@ -9,11 +9,11 @@ from pydantic import Field, ValidationInfo, field_validator, model_validator
 from typing_extensions import Annotated
 
 from aind_data_schema.base import AindGeneric, AindGenericType, AindModel, AwareDatetimeWithDefault
-from aind_data_schema.models.coordinates import RelativePosition, Size3d
+from aind_data_schema.core.coordinates import RelativePosition, Size3d
 from aind_data_schema.models.harp_types import HarpDeviceType
 from aind_data_schema.models.harp_types import Olfactometer as OlfactometerHarpType
 from aind_data_schema.models.organizations import InteruniversityMicroelectronicsCenter, Organization
-from aind_data_schema.models.reagent import Reagent
+from aind_data_schema.core.reagent import Reagent
 from aind_data_schema.models.units import FrequencyUnit, PowerUnit, SizeUnit, SpeedUnit, TemperatureUnit, UnitlessUnit
