+++ src/aind_data_schema/core/session.py
@@ -11,18 +11,18 @@ from typing_extensions import Annotated
 
 from aind_data_schema.base import AindCoreModel, AindGeneric, AindGenericType, AindModel, AwareDatetimeWithDefault
 from aind_data_schema.core.procedures import Anaesthetic
-from aind_data_schema.imaging.tile import Channel
-from aind_data_schema.models.coordinates import (
+from aind_data_schema.core.tile import Channel
+from aind_data_schema.core.coordinates import (
     CcfCoords,
     Coordinates3d,
     Rotation3dTransform,
     Scale3dTransform,
     Translation3dTransform,
 )
-from aind_data_schema.models.devices import Calibration, Maintenance, RelativePosition, Scanner, Software, SpoutSide
+from aind_data_schema.core.devices import Calibration, Maintenance, RelativePosition, Scanner, Software, SpoutSide
 from aind_data_schema.models.modalities import Modality
 from aind_data_schema.models.process_names import ProcessName
-from aind_data_schema.models.stimulus import (
+from aind_data_schema.core.stimulus import (
     AuditoryStimulation,
     OlfactoryStimulation,
     OptoStimulation,
