+++ src/aind_data_schema/core/stimulus.py
