--- src/aind_data_schema/imaging/__init__.py
@@ -1,1 +1,0 @@
-""" imports for imaging subpackage """
